test : Car 도메인 테스트 코드 추가
