+++ src/test/kotlin/controller/RacingGameControllerTest.kt
@@ -1,10 +1,9 @@
 package controller
 
-import model.Car
-import model.GameResult
+import domain.Car
+import domain.GameResult
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.AfterEach
-import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import view.InputView
@@ -12,7 +11,6 @@ import view.OutputView
 import java.io.ByteArrayOutputStream
 import java.io.OutputStream
 import java.io.PrintStream
-import java.util.Random
 
 class RacingGameControllerTest {
     private lateinit var outputStream: OutputStream
@@ -38,22 +36,6 @@ class RacingGameControllerTest {
     private val cars: List<Car> = listOf(hwanCar, siaCar)
 
     private val raceController = RacingGameController(InputView(), OutputView())
-
-    @Test
-    fun `자동차가 1칸 이동하면 현재 위치가 1 늘어난다`() {
-        val random = Random(1234)
-        val position = hwanCar.currentPosition
-        hwanCar.moveCar(random)
-        assertEquals(position + 1, hwanCar.currentPosition)
-    }
-
-    @Test
-    fun `랜덤 숫자가 4 미만인 경우, 자동차의 위치는 바뀌지 않는다`() {
-        val random = Random(123)
-        val position = siaCar.currentPosition
-        siaCar.moveCar(random)
-        assertEquals(position, siaCar.currentPosition)
-    }
 
     @Test
     fun `가장 많이 움직인 자동차가 우승한다`() {
